+++ tests/views/test_activation.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from urllib import quote
 
+import pytest
 import transaction
 
 from pyramid_fullauth.models import User
@@ -48,6 +49,9 @@ def test_account_activation_key_with_trash_chars(user, db_session, default_app):
     '''
         Register:Activate user
     '''
+    if db_session.connection().dialect.name == 'mysql':
+        pytest.xfail("failing due mysql default character set being latin1 and collation latin1_swedish_ci")
+
     user = db_session.merge(user)
 
     activate_key = user.activate_key