+++ CHANGES.rst
@@ -24,6 +24,8 @@ tests
     - pep8
     - pyflakes
     - mccabe
+- add pytest-dbfixtures, and run tests against postgresql and mysql as well
+- drop python 2.6 from tests
 
 
 0.2.3